My version so far
Ice platforms move and bob, planks fall when walked across with the middle ones having a delay to destroy themselves so the player can actually hit the spikes. Simple AI added to the background of one scene, and one prototype enemy on the dock however he seems to get stuck. Still working on the AI