CharacterBP,EnemyBP,Weapons,PickUps
The character has an Axe and the Animation Blueprint for the character is done
The enemy has an attached sword to his hand. Please note the sword doesnt spawn in the scene like the character weapon does.
HealthPickUp is created and can be easily spawn in the scene by using the HealthPickUp Actor. There is also an animation after the health is picked but needs to be setted up.